refactor: Move mock file to mocks package
