+++ providers/multi-provider/internal/mocks/provider_mock.go
@@ -7,7 +7,7 @@
 //
 
 // Package strategies is a generated GoMock package.
-package strategies
+package mocks
 
 import (
 	context "context"
