+++ providers/multi-provider/internal/strategies/first_match_test.go
@@ -3,17 +3,18 @@ package strategies
 import (
 	"context"
 	"fmt"
+	"github.com/open-feature/go-sdk-contrib/providers/multi-provider/internal/mocks"
 	of "github.com/open-feature/go-sdk/openfeature"
 	"github.com/stretchr/testify/assert"
 	"go.uber.org/mock/gomock"
 	"testing"
 )
 
-func createMockProviders(ctrl *gomock.Controller, count int) ([]*NamedProvider, map[string]*MockFeatureProvider) {
+func createMockProviders(ctrl *gomock.Controller, count int) ([]*NamedProvider, map[string]*mocks.MockFeatureProvider) {
 	providers := make([]*NamedProvider, 0, count)
-	providerMocks := make(map[string]*MockFeatureProvider)
+	providerMocks := make(map[string]*mocks.MockFeatureProvider)
 	for index := range count {
-		provider := NewMockFeatureProvider(ctrl)
+		provider := mocks.NewMockFeatureProvider(ctrl)
 		namedProvider := NamedProvider{
 			Provider: provider,
 			Name:     fmt.Sprintf("%d", index),
